// FrontendProyectoAngular/src/app/models/empleados.ts
export class createEmployee{
    NOMBRE: string = "";
    APELLIDOS: string = "";
    UBICACION: string = "";
    PUESTO: string = "";
    NUMERO_TELEFONO: string = "";
    EXTENSION: string = "";
}

export class deleteEmployee{
    ID_EMPLEADO: number = 0;
    NOMBRE: string = "";
    APELLIDOS: string = "";
    UBICACION: string = "";
    PUESTO: string = "";
    NUMERO_TELEFONO: string = "";
    EXTENSION: string = "";
}